+++ src/main/java/racingcar/controller/PrintController.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.controller;
 
 import racingcar.unit.Cars;
